backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
datasets==2.14.6
scikit-learn==1.3.2
nltk==3.8.1
numpy==1.24.3
pandas==2.0.3
networkx==3.2
pdfplumber==0.9.0
python-docx==0.8.11
protobuf==4.25.1